moved various event components into events dir
