+++ TasteBuddiesClient/src/app/events/event.component.html
