+++ TasteBuddiesClient/src/app/events/event.component.spec.ts
