+++ TasteBuddiesClient/src/app/events/event.component.ts
