+++ TasteBuddiesClient/src/app/events/user-events/user-events.component.ts
